% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprData.R
\name{exprData}
\alias{exprData}
\title{Expression data subset}
\usage{
exprData(value = NULL, pvalue = 0.1, ds = NA, shrinkLFC = FALSE, calcDE = TRUE)
}
\arguments{
\item{value}{Numeric. LFC threshold to subset dataset based on DESeq2 results values. Taken as absolute value.}

\item{pvalue}{Numeric. P-value threshhold.}

\item{ds}{DESeq2 dataset, output of \code{DESeq(x)}.}

\item{shrinkLFC}{Logical. Apply \code{lfcShrink} function to generate DESeq2 results.}

\item{calcDE}{Logical. Calcilate differential expression by substracting 'control' from 'expreriment' columns. To work this option requires \code{colData(dds)} to contain column called 'clustering'. This column must tag samples as 'control' or 'experiment'.}
}
\description{
\code{exprData}Returns a list made of count matrix and DESeq2 results data.frame.
}
