% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vic3_boxplot}
\alias{vic3_boxplot}
\title{Sample variance box plot}
\usage{
vic3_boxplot(counts, ntop = 500)
}
\arguments{
\item{counts}{Counts table output \code{assay(x)} or matrix.}

\item{ntop, }{numeric Number of top genes with most variance.}
}
\description{
\code{vic3_boxplot} returns box plot diagram of varince per column in counts matrix.
}
\details{
The function accepts matrix or \code{assay(x)} count table.
}
\examples{

se <- readRDS(system.file("extdata", "se_vic3_2020.RData", package = "vic3PCD"))

vic3_boxplot(assay(se))
}
