% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vic3_plotPCA}
\alias{vic3_plotPCA}
\title{PCA plot}
\usage{
vic3_plotPCA(counts, group_vector, ntop = 500, returnData = FALSE)
}
\arguments{
\item{counts}{Counts table output \code{assay(x)} or matrix.}

\item{group_vector}{Column name in DESeq2 \code{colData(x)}.}

\item{ntop}{Numeric. Number of top genes with most variance.}

\item{returnData}{Logical (default FALSE). Return PC1 and PC2 data.frame.}
}
\description{
\code{vic3_plotPCA} returns PCA plot for normalized counts output by DESeq2.
}
\details{
The function works similarly to DESeq2 \code{plotPCA}. It accepts count tables which could be accessed with \code{assay} function.
}
\examples{

## Load SE dataset
se <- readRDS(system.file("extdata", "se_vic3_2020.RData", package = "vic3PCD"))

## DESeq dataset
dds <- DESeqDataSet(se, design = ~ condition)

## DESeq analysis
dds <- DESeq(dds)
## Filter genes with more than 10 aligned reads
keep <- rowSums(counts(dds)) >= 10
dds <- dds[keep,]
## Normised counts
rld <- rlog(dds, blind = F)

## Manual color
clr = c("dodgerblue3", "coral3")
names(clr) <- unique(colData(rld)[,"condition"])

vic3_plotPCA(counts = assay(rld), group_vector = colData(rld)[,"condition"]) +
scale_colour_manual(values=clr, name = "Condition:") +
geom_text(label = colData(rld)[,"exset"], size = 3, colour = "black", fontface = "italic", hjust = 0, nudge_x = 3) +
xlim(-20,60) +
ylim(-25,25)
}
